Get sub info through phone id in the extras.
In the origin design, it used sub id from extras to get sub
info. But sub id was invalid in sometimes, hence the carrier
config did not been cached properly.

To fix it, using phone id to get sub info. If remove sub or
sub info is not ready, clean the cached carrier config of the
sub.

Change-Id: I2984a0437cd85ac5c8741fa386f69b774d4f7f88
CRs-Fixed: 2352445